Struct glium_shapes::sphere::SphereBuilder

pub struct SphereBuilder { /* fields omitted */ }

Responsible for building and returning a Sphere object.

By default, the sphere is defined as a unit-sphere (e.g. a radius of 1) with its centre-of-mass located at the origin. This can be overriden using the transformation methods on this object.

The resultant geometry is constructed to suit OpenGL defaults - assuming a right-handed coordinate system, front-facing polygons are defined in counter-clock-wise order. Vertex normals point in the direction of their respective face (such that the shape appears faceted when lit). Vertex texture coordinates define a spherical-projection on the object.

Methods

impl SphereBuilder

fn new() -> SphereBuilder

Create a new SphereBuilder object.

fn with_divisions(self, u: usize, v: usize) -> Self

Specify the number of divisions to make in the u direction (horizontal), and v direction (vertical). By default, the builder will use 12 divisions in both axes.

fn scale(self, x: f32, y: f32, z: f32) -> Self

Apply a scaling transformation to the shape.

The scale, translate, and rotate functions accumulate, and are not commutative. The transformation functions are intended to provide flexibility in model-space. For per-instance world-space transformations, one should prefer to share as few shapes as possible across multiple instances, and instead rely on uniform constants in the shader and/or instanced drawing.

fn translate(self, x: f32, y: f32, z: f32) -> Self

Apply a translation transformation to the shape.

The scale, translate, and rotate functions accumulate, and are not commutative. The transformation functions are intended to provide flexibility in model-space. For per-instance world-space transformations, one should prefer to share as few shapes as possible across multiple instances, and instead rely on uniform constants in the shader and/or instanced drawing.

fn rotate_x(self, radians: f32) -> Self

Apply a rotation transformation to the shape about the x-axis.

The scale, translate, and rotate functions accumulate, and are not commutative. The transformation functions are intended to provide flexibility in model-space. For per-instance world-space transformations, one should prefer to share as few shapes as possible across multiple instances, and instead rely on uniform constants in the shader and/or instanced drawing.

fn rotate_y(self, radians: f32) -> Self

Apply a rotation transformation to the shape about the y-axis.

The scale, translate, and rotate functions accumulate, and are not commutative. The transformation functions are intended to provide flexibility in model-space. For per-instance world-space transformations, one should prefer to share as few shapes as possible across multiple instances, and instead rely on uniform constants in the shader and/or instanced drawing.

fn rotate_z(self, radians: f32) -> Self

Apply a rotation transformation to the shape about the z-axis.

The scale, translate, and rotate functions accumulate, and are not commutative. The transformation functions are intended to provide flexibility in model-space. For per-instance world-space transformations, one should prefer to share as few shapes as possible across multiple instances, and instead rely on uniform constants in the shader and/or instanced drawing.

fn build<F>(self, display: &F) -> Result<Sphere, ShapeCreationError> where F: Facade

Build a new Sphere object.

fn build_vertices(&self) -> Result<Vec<Vertex>, ShapeCreationError>

Build the shape vertices and return them in a vector.

Useful if you wish to do other things with the vertices besides constructing a Sphere object (e.g. unit testing, further processing, etc).

fn num_caps(&self) -> usize

Returns the number of caps in the resultant sphere geometry. The current implementation will always return 2.

fn num_vertices_per_cap_face(&self) -> usize

Returns the number of vertices generated for each cap face. The current implementation will always return 3.

fn num_vertices_per_cap(&self) -> usize

Returns the total number of vertices in each cap.

fn num_slices(&self) -> usize

Returns the number of vertical slices in the resultant sphere geometry. The resultant value will depend on the number of v divisions specified on the builder.

fn num_vertices_per_slice_face(&self) -> usize

Returns the total number of vertices in each vertical slice face (e.g. excluding caps). The current implementation will always return 6.

fn num_vertices_per_slice(&self) -> usize

Returns the total number of vertices in each vertical slice.

fn num_vertices(&self) -> usize

Returns the total number of vertices that will be generated by the builder.